Beta EOCore build: 2.3.205
Beta EOCore Change Notes:
Bug Fixes

- Update syntax due to swift issue with NSNumber during json creation.

Improvements:

- New minimal os is 12.0 from 9.0 for library.
- Add more debug information what is being used for configuration settings.

Our environment for this release:

- Xcode 13.2.1
- MacOS 11.6.1
- iOS 12.x to 15.x
- Supported architectures:
	- simulator
		- arm64
		- x86_64
	- device
		- arm64

Known Issues

- After iOS 13, library no longer supports iPad because of Apple changes in iPad OS.
- Dual sim features are in beta. Example information using multiple carriers.

+++ SDKs/iOS/Debug/EOCore.xcframework/ios-arm64_armv7/EOCore.framework/Headers/EOCore.h
@@ -128,11 +128,11 @@
  * @return If the configuration data was successfully updated or not.
  */
 - (BOOL)updateConfigForModuleName:(NSString *)moduleName withData:(NSDictionary *)newConfig;
-- (double)getCurrentSessionStartTimeforModuleName:(NSString *)moduleName;
--(NSString *)getConfigItemAsString:(NSString *)key forModuleName:(NSString *)moduleName;
--(NSNumber *)getConfigItemAsNumber:(NSString *)key forModuleName:(NSString *)moduleName;
--(BOOL ) getConfigItemAsBool:(NSString *)key withDefault:(BOOL)theDefault forModuleName:(NSString *)moduleName;
--(NSNumber *)defaultLogLevel;
+- (unsigned long long)getCurrentSessionStartTimeforModuleName:(NSString *)moduleName;
+- (NSString *)getConfigItemAsString:(NSString *)key forModuleName:(NSString *)moduleName;
+- (NSNumber *)getConfigItemAsNumber:(NSString *)key forModuleName:(NSString *)moduleName;
+- (BOOL ) getConfigItemAsBool:(NSString *)key withDefault:(BOOL)theDefault forModuleName:(NSString *)moduleName;
+- (NSNumber *)defaultLogLevel;
 
 /*!
  * @abstract Generate device ID for Session
